Updates
Restructured and organized the Updates view component.
