+++ src/components/index.jsx
@@ -5,6 +5,10 @@ import FirstSectionNews from './homeComponents/sectionNews/firstSectionNews';
 import SecondSectionNews from './homeComponents/sectionNews/secondSectionNews';
 import FrequentQuestions from './homeComponents/frequentQuestions/frequentQuestions';
 import BannerHome from './homeComponents/bannerHome/bannerHome';
+import { BannerUpdates, BannerUpdatesMobile } from './updatesComponents/banner/bannerUpdates';
+import { TitleUpdates, TitleUpdatesMobile } from './updatesComponents/title/titleUpdates';
+import UpdatesSections from './updatesComponents/updatesSections/updatesSections';
+import DownloadFiles from './updatesComponents/updatesDownloadFiles/updatesDownloadFiles';
 
 export {
     Navbar,
@@ -13,5 +17,11 @@ export {
     FirstSectionNews,
     SecondSectionNews,
     FrequentQuestions,
+    BannerUpdates,
+    BannerUpdatesMobile,
+    TitleUpdates,
+    TitleUpdatesMobile,
+    UpdatesSections,
+    DownloadFiles,
     Footer
 }
